requirements.txt
bcrypt>=4.3.0
google-genai>=1.38.0
google-generativeai>=0.8.5
pandas>=2.3.2
pdfplumber>=0.11.7
plotly>=6.3.0
pypdf2>=3.0.1
python-dotenv>=1.1.1
sift-stack-py>=0.9.1
streamlit>=1.49.1
